+++ clomonitor-core/Cargo.toml
@@ -8,5 +8,4 @@ anyhow = "1.0.52"
 askalono = "0.4.4"
 glob = "0.3.0"
 regex = "1.5.4"
-thiserror = "1.0.30"
 serde = { version = "1.0", features = ["derive"] }
