+++ clomonitor-core/src/linter/mod.rs
@@ -1,9 +1,26 @@
 mod check;
 
-use anyhow::Error;
+use anyhow::{format_err, Error};
 use check::Globs;
 use serde::{Deserialize, Serialize};
 use std::path::Path;
+
+/// Supported linters.
+#[derive(Debug)]
+pub enum Linter {
+    Core = 0,
+}
+
+impl std::convert::TryFrom<i32> for Linter {
+    type Error = Error;
+
+    fn try_from(linter_id: i32) -> Result<Self, Self::Error> {
+        match linter_id {
+            0 => Ok(Linter::Core),
+            _ => Err(format_err!("invalid linter id")),
+        }
+    }
+}
 
 /// A linter report.
 #[derive(Debug, Serialize, Deserialize)]
