+++ clomonitor-tracker/src/main.rs
@@ -1,16 +1,13 @@
+mod repository;
 mod tracker;
 
-use crate::tracker::*;
 use anyhow::{format_err, Error};
 use clap::Parser;
 use config::{Config, File};
 use deadpool_postgres::{Config as DbConfig, Runtime};
-use futures::future;
-use futures::stream::{FuturesUnordered, StreamExt};
 use openssl::ssl::{SslConnector, SslMethod, SslVerifyMode};
 use postgres_openssl::MakeTlsConnector;
 use std::path::PathBuf;
-use tracing::{error, info};
 use which::which;
 
 #[derive(Debug, Parser)]
@@ -36,8 +33,6 @@ async fn main() -> Result<(), Error> {
         return Err(format_err!("git not found in PATH"));
     }
 
-    info!("tracker started");
-
     // Setup configuration
     let mut cfg = Config::new();
     cfg.set_default("db.dbname", "clomonitor")?;
@@ -51,31 +46,8 @@ async fn main() -> Result<(), Error> {
     let db_cfg: DbConfig = cfg.get("db").unwrap();
     let db_pool = db_cfg.create_pool(Some(Runtime::Tokio1), connector)?;
 
-    // Get repositories to process
-    let repositories = get_repositories(db_pool.get().await?).await?;
-    if repositories.is_empty() {
-        info!("no repositories found");
-        info!("tracker finished");
-        return Ok(());
-    }
+    // Run tracker
+    tracker::run(cfg, db_pool).await?;
 
-    // Process repositories
-    info!("processing repositories");
-    let mut futs = FuturesUnordered::new();
-    for repo in repositories {
-        let db = db_pool.get().await?;
-        futs.push(tokio::spawn(async move {
-            let repository_id = repo.id();
-            if let Err(err) = process_repository(db, repo).await {
-                error!("error processing repository {repository_id}: {err}");
-            }
-        }));
-        if futs.len() == cfg.get::<usize>("tracker.concurrency").unwrap() {
-            futs.next().await;
-        }
-    }
-    future::join_all(futs).await;
-
-    info!("tracker finished");
     Ok(())
 }
